Struct futures_test::harness::Harness

pub struct Harness<T> { /* fields omitted */ }

Wraps a future, providing an API to interact with it while off task.

This wrapper is intended to use from the context of tests. The future is effectively wrapped by a task and this harness tracks received notifications as well as provides APIs to perform non-blocking polling as well blocking polling with or without timeout.

Methods

impl<T> Harness<T>

pub fn new(obj: T) -> Self

Wraps obj in a test harness, enabling interacting with the future while not on a Task.

pub fn with<F, R>(&mut self, f: F) -> R where
F: FnOnce(&mut Self) -> R,

pub fn is_notified(&self) -> bool

Returns true if the inner future has received a readiness notification since the last action has been performed.

pub fn get_ref(&self) -> &T

Returns a reference to the inner future.

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the inner future.

pub fn into_inner(self) -> T

Consumes self, returning the inner future.

impl<F, T, E> Harness<PollFn<F>> where
F: FnMut() -> Poll<T, E>,

pub fn poll_fn(f: F) -> Self

Wraps the poll_fn in a harness.

impl<T: Future> Harness<T>

pub fn poll(&mut self) -> Poll<T::Item, T::Error>

Polls the inner future.

This function returns immediately. If the inner future is not currently ready, NotReady is returned. Readiness notifications are tracked and can be queried using is_notified.

pub fn wait(&mut self) -> Result<T::Item, T::Error>

Waits for the internal future to complete, blocking this thread's execution until it does.

pub fn wait_timeout(
&mut self,
dur: Duration
) -> Result<T::Item, TimeoutError<T::Error>>

Waits for the internal future to complete, blocking this thread's execution for at most dur.

impl<T: Stream> Harness<T>

pub fn poll_next(&mut self) -> Poll<Option<T::Item>, T::Error>

Polls the inner future.

This function returns immediately. If the inner future is not currently ready, NotReady is returned. Readiness notifications are tracked and can be queried using is_notified.