Struct futures_test::io::read::Limited

pub struct Limited<Io> { /* fields omitted */ }

I/O wrapper that limits the number of bytes written or read on each call.

See the limited and limited_write methods.

Methods

impl<Io> Limited<Io>

pub fn get_ref(&self) -> &Io

Acquires a reference to the underlying I/O object that this adaptor is wrapping.

pub fn get_mut(&mut self) -> &mut Io

Acquires a mutable reference to the underlying I/O object that this adaptor is wrapping.

pub fn get_pin_mut<'a>(self: Pin<&'a mut Self>) -> Pin<&'a mut Io>

Acquires a pinned mutable reference to the underlying I/O object that this adaptor is wrapping.

pub fn into_inner(self) -> Io

Consumes this adaptor returning the underlying I/O object.