Struct futures_test::io::write::TrackClosed

pub struct TrackClosed<T> { /* fields omitted */ }

Async wrapper that tracks whether it has been closed.

See the track_closed methods on:

SinkTestExt

AsyncWriteTestExt

Implementations

impl<T> TrackClosed<T>

pub fn is_closed(&self) -> bool

Check whether this object has been closed.

pub fn get_ref(&self) -> &T

Acquires a reference to the underlying object that this adaptor is wrapping.

pub fn get_mut(&mut self) -> &mut T

Acquires a mutable reference to the underlying object that this adaptor is wrapping.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut T>

Acquires a pinned mutable reference to the underlying object that this adaptor is wrapping.

pub fn into_inner(self) -> T

Consumes this adaptor returning the underlying object.